# 8297359: RISC-V: improve performance of floating Max Min intrinsics