@VladimirKempik 2022-11-25
Thanks for looking at it.
Tier1 is fine, running the rest